Version 41 Succeed Ges Mode Succeed next pod install Realm

+++ TeamBook/Controller/TableControllerVocal.swift
@@ -127,41 +127,48 @@ extension TableControllerVocal {
     }
     
     @objc func handleLongPress(_ gestureRecognizer: UILongPressGestureRecognizer) {
-        selectCellEmail = 0
-        let point = gestureRecognizer.location(in: self.tableView)
-        if let indexPath = self.tableView.indexPathForRow(at: point) {
-            selectCellEmail = indexPath.row
-            if gestureRecognizer.state == .began {
-                let collectionRef = db.collection("collection:\(buttonName)")
-                collectionRef.getDocuments { (querySnapshot, error) in
-                    if let documents = querySnapshot?.documents{
-                        for doc in documents{
-                            if let selfIntroduction = doc.data()["SelfIntroduction"]as? String{
-                                if self.user!.email == self.firebaseDataArray[self.selectCellEmail].allEmailText{
-                                    let alertController = UIAlertController(title: "Delete Item",message: "Are you sure you want to delete this item?",preferredStyle: .alert)
-                                    let deleteAction = UIAlertAction(title: "Delete", style: .destructive) { _ in
-                                        let ToDelete = selfIntroduction
-                                        self.deleteFromFirebase(SelfIntroduction: ToDelete)
-                                        self.firebaseDataArray.remove(at: indexPath.row)
-                                        print(ToDelete)
-                                        self.tableView.reloadData()
+        if user != nil {
+            selectCellEmail = 0
+            let point = gestureRecognizer.location(in: self.tableView)
+            if let indexPath = self.tableView.indexPathForRow(at: point) {
+                selectCellEmail = indexPath.row
+                if gestureRecognizer.state == .began {
+                    let collectionRef = db.collection("collection:\(buttonName)")
+                    collectionRef.getDocuments { (querySnapshot, error) in
+                        if let documents = querySnapshot?.documents{
+                            for doc in documents{
+                                if let selfIntroduction = doc.data()["SelfIntroduction"]as? String{
+                                    if self.user!.email == self.firebaseDataArray[self.selectCellEmail].allEmailText{
+                                        let alertController = UIAlertController(title: "Delete Item",message: "Are you sure you want to delete this item?",preferredStyle: .alert)
+                                        let deleteAction = UIAlertAction(title: "Delete", style: .destructive) { _ in
+                                            let ToDelete = selfIntroduction
+                                            self.deleteFromFirebase(SelfIntroduction: ToDelete)
+                                            self.firebaseDataArray.remove(at: indexPath.row)
+                                            print(ToDelete)
+                                            self.tableView.reloadData()
+                                        }
+                                        let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
+                                        alertController.addAction(deleteAction)
+                                        alertController.addAction(cancelAction)
+                                        self.present(alertController, animated: true, completion: nil)
                                     }
-                                    let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
-                                    alertController.addAction(deleteAction)
-                                    alertController.addAction(cancelAction)
-                                    self.present(alertController, animated: true, completion: nil)
-                                }
-                                else{
-                                    let alertController = UIAlertController(title: "What?!", message: "You want to delete others message?", preferredStyle: .alert)
-                                    let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
-                                    alertController.addAction(cancelAction)
-                                    self.present(alertController,animated: true,completion: nil)
+                                    else{
+                                        let alertController = UIAlertController(title: "What?!", message: "You want to delete others message?", preferredStyle: .alert)
+                                        let cancelAction = UIAlertAction(title: "Cancel", style: .cancel, handler: nil)
+                                        alertController.addAction(cancelAction)
+                                        self.present(alertController,animated: true,completion: nil)
+                                    }
                                 }
                             }
                         }
                     }
                 }
             }
+        }else{
+            let alertController = UIAlertController(title: "Please Log In!",message: "If U want to send message,plz sign up and log in",preferredStyle: .alert)
+            let cancelAction = UIAlertAction(title: "OK", style: .cancel, handler: nil)
+            alertController.addAction(cancelAction)
+            self.present(alertController, animated: true, completion: nil)
         }
     }
 }
